atoi

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -18,5 +18,5 @@ int _atoi(char *s)
 		else if (num > 0)
 			break;
 	}
-	return num * sign;
+	return (int)(num * sign);
 }
